Set linux/amd64 platform for localnet Makefile targets
This allows to avoid platform-related warnings on M1 Apple developer machines.

Please note that the client image for localnet is built for linux/amd64 platform
per the `networks/local/Makefile` and we need to preserve the platform used.

I tried to approach this problem setting DOCKER_DEFAULT_PLATFORM instead of
--platform linux/amd64 for each individual docker call. However, given line in
a Make recipe is executed using a separate shell invocation, even exported
environment variable was not propagated to docker calls.

Exporting variable using Make's construct could be too much given we want this
platform to be set explicitly only for localnet builds. (Maybe ? the build-linux
target as well but it is out of the scope of the current work.)