Use evmos_9000-10 chainId for localnet
The logic in `x/evm/keeper.keeper.go` does not allow to start the client
with other chain ID than 9000 or 9001.

evmos_9001-2 is a chain ID of Evmos 2 mainnet.
evmos_9000-4 is a chain ID of Evmos Public Testnet that is currently
active.

We use evmos_9000-10 for locanet. This value shouldn't be used
elsewhere.