Change the directory in the permission issue fix step
The `semgrep.yml` was sourced from the `evmoss/evmos` GitHub project. We now
need to adjust it to work with our repo's config. The previous config was
resulting in the following error:

```
git remote add get-diff-action
  >> fatal: detected dubious ownership in repository at '/__w/meso/meso'
  >> To add an exception for this directory, call:
  >>
  >> 	git config --global --add safe.directory /__w/meso/meso
```

This is an error that is related to a Git's vulnerability fix (introduced in Git
v2.35.2).

> This vulnerability affects users working on multi-user machines where a
> malicious actor could create a .git directory in a shared location above a
> victim’s current working directory. On Windows, for example, an attacker could
> create C:\.git\config, which would cause all git invocations that occur
> outside of a repository to read its configured values.

[source](https://github.blog/2022-04-12-git-security-vulnerability-announced/#cve-2022-24765)