Comment out the step running the dependency review
The
[`actions/dependency-review-action@v3`](https://github.com/actions/dependency-review-action/tree/v3/)
action used in the workflow for detecting vulnerabilities and invalid licenses
is not available for private repositories that don't have GitHub Advanced
Security licensed. That's why we were getting `Dependency review is not
supported on this repository. Please ensure that Dependency graph is enabled
along with GitHub Advanced Security on private repositories, see
https://github.com/thesis/meso/settings/security_analysis` error. Until we
change the repo to public or enable GitHub Advanced Security, we're commenting
out the step that runs this action.