Remove the Semgrep workflow
The Evmos project used [Semgrep](https://semgrep.dev/) tool to identify security
vulnerabilities and ensure the code quality. The tool was integrated with the CI
in the `semgrep.yml` file.

We decided to not use the tool for now ('_It's a cool tool, but without someone
to own its maintenance I don't think it's going to give us the value it's
designed for._'). But we want to revisit this decision in the future (we're
tracking it under https://github.com/thesis/meso/issues/33).

If we were to integrate the tool into our project in the future, these are some
decisions and actions that would need to be done:

* Choose a [Semgrep plan](https://semgrep.dev/pricing/).
* [Sign in using GH account and create an organization](https://semgrep.dev/docs/semgrep-cloud-platform/getting-started/#signing-in-to-semgrep-cloud-platform)
  (from what I understand, the account used to sign in would be consider an
  admin of the Semgrep Cloud Platform).
* [Configure `SEMGREP_APP_TOKEN`](https://semgrep.dev/docs/semgrep-ci/running-semgrep-ci-with-semgrep-cloud-platform/#creating-a-semgrep_app_token).
* Set `SEMGREP_APP_TOKEN` as a GH secret.
* Add the CI workflow responsible for running the tool.