Remove the `cleanup-runs` job
The job was configured so that it would cancel previously running instances of
the workflow run on the same branch. The job had strange configuration (the
conditions in the `if` clause  were not making much sense in the context of the
`pull_request` workflow trigger). What's more, action used in the job is marked
as deprecated (because it's functionality is now natively supported by GH
Actions).
We're removing the job and instead we'll be using it's GH's equivalent -
the
[`concurrency`](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#concurrency)
command.