Comment out the `goreleaser.yml` project
The workflow, in conjunction with the Makefile config, was used by Evmos to
automate the process of creating and publishing releases for the
`github.com/evmos/evmos` Go project. We need to decide if we want to use similar
release config for our project as the one used for `github.com/evmos/evmos` and
adjust the workflow and Makefile config accordingly.