Remove the `docker-push.yml` workflow
The workflow was copied from the `evmos/evmos` repository and was supposed to
push new docker images to DockerHub when a new tag with version is pushed.

We're removing the workflow as we don't wan to publish Dosker images for now. If
we will want to push Docker images to DockerHub in the future, we may use this
workflow as a template. But for now we don't need it.