Expose JSON-RPC ports of localnet clients
This change makes every client of the localnet to expose JSON-RPC over
HTTP and WebSocket. This allows to interact with the network from the
host machine.

It is still unclear to me what the difference is between
`cmd/evmosd/testnet.go` and `client/testnet.go`. The two files are
almost identical and expose the same `evmosd testnet init-files` CLI
command. For the time being, I decided to keep both files and mirror
their content. We will potentially deal with the duplication
separately.