Setup internal testnet faucet
Here we introduce an account for internal testnet faucet. The account
is funded with 1000 BTC printed out of the air. We are also taking
an opportunity and increase liquid balance for existing validators.
From now on, they all have 1000 BTC of liquid balance from which 100 BTC is
staked.

The above requires changes in the chain's genesis file so, we were
forced to generate new testnet artifacts and reset the existing internal
testnet chain. The keystore of the `mezo-faucet` account was loaded
to the Kubernetes cluster as `mezo-faucet-keystore` secret. We can
leverage it to expose a public faucet service in the future.