Remove the Cosmos `x/authz` module
The `x/authz` manages authorization grants. A grant is an allowance to execute
a Cosmos `Msg` by the grantee on behalf of the granter. In Mezo, we don't
necessarily need this capability as all user-facing actions
will go through EVM. That said, authorization grants should live on
the EVM contract level, rather than being managed by the `x/authz` module
on the Cosmos level.