Remove Evmos `x/erc20` module
This module is used to create mappings between ERC20 tokens on EVM and
Cosmos denoms in the `x/bank` module. This is not needed for Mezo.
Users will exclusively use EVM transactions to interact with the chain so there
is no sense to map custom ERC20 to the Cosmos runtime. The only exception
is the gas token (BTC). However, this mapping will be handled through an
EVM precompile interacting with the `x/bank` module directly. Moreover, if
there is a need to map EVM ERC20 to `x/bank` in the future, precompiles
should be preferred anyway as their provide better UX than `x/erc20`.
Last but not least, we need to remove `x/erc20` as it depends on IBC which
will be removed as well.