Comment out upgrade handlers
The decision what to do with upgrades is not clear for now.
A lot of them depends on already removed modules and cause compilation
errors beyond `app.go` (those errors are expected at this
stage of the cleanup). The best way to address it right now is to
comment out upgrade handlers