Integrate `x/poa` with `x/evm`
The `x/evm` expects to interact with a staking keeper exposing two
functions:
- GetValidatorByConsAddr
- GetHistoricalInfo

Our `x/poa` keeper needs to implement both to comply with that interface.
The first one is trivial as the `x/poa` module allows fetching
validators by consensus address.

The second function is more tricky. It is used by `x/evm` module
to resolve hashes of blocks that are behind the chain tip. In order to
make it work, we need to port the historical info tracking feature from
the `x/staking` Cosmos module to our `x/poa` module. This will be handled
separately. For now, we are just providing a stub implementation of that
function.