Configure Docker linter to ignore apt-get non-pinned versions
The `super-linter` tool used as part of the `run-lint` CI jobs
triggers `hadolint` to lint Dockerfiles. This tool raises an error
if `apt-get install` command installs a package without a pinned version
which is the case for all Dockerfiles in the Mezo codebase. To overcome
that problem, we make `hadolint` to ignore rule `DL3008` which is
responsible for that check.