RFC-2: Bridging Bitcoin to Mezo
Mezo chain uses Bitcoin as a base token. The EVM-compatible version of Bitcoin
is tBTC and the tBTC Bridge will be used to bring Bitcoin to Mezo. Currently,
the tBTC Bridge ledger exists on Ethereum, so Ethereum-to-Mezo tBTC bridging has
to be achieved in the first release. The bridging mechanism has to be compatible
with the Bitcoin token precompile on Mezo and reflect the balances in the x/bank
module.

This RFC describes bridging tBTC from Ethereum to Mezo. Bridging back and
bridging other assets is out of the scope and will be covered in separate RFC
documents. Also, the potential future work of moving the tBTC Bridge ledger
from Ethereum to Mezo is out of the scope of this RFC.