RFC-2: Reworked the concept of bridge validators
From now on, all validators in the network inspect Ethereum sidecars to
confirm that what bridge validators propose makes sense.

This model ensures that only the validators running the full Ethereum
nodes can initiate bridging and the entire network monitors the initiated
bridging operations. In the worst case of a single compromised JSON-RPC
endpoint being used by the rest of the network, the bridging halts but
no artificial Bitcoins are minted.

In reality, the rest of the network will not use the same JSON-RPC endpoint
and even if the three most popular JSON-RPC providers are used, the entire
network performs the verification without blindly trusting a subset of
validators. Also, this solution is easier to extend if we decide to
integrate ETH2 light clients into sidecars (see Future Work section).