RFC-2: Warn about Ethereum JSON-RPC decentralization
Since all validators in the network need to run Ethereum sidecars blocking
the most popular JSON-RPC providers is not feasible. We will still require
bridge validators to run full Ethereum nodes (via social contract) but we
can warn in logs on the sidecar start about helping to decentralize the
network if we detect Infura or Alchemy endpoint being used.