Replace nodes IDs
Here we use the outcome of the adjusted `script/public-testnet.sh` script
and replace the testnet nodes IDs stored on the Kubernetes cluster
with fresh ones. This is needed because the script generated new nodes
identities that are now referenced in the updated genesis file. We cannot
just reuse the existing ones.