RFC-3: Bridging non-Bitcoin assets to Mezo
RFC-2: Bridging Bitcoin to Mezo describes the mechanism of bridging
Bitcoin represented on EVM as tBTC to Mezo chain. Bitcoin is the base
asset of Mezo and requires a separate bridging path that may in the
future be transformed into direct tBTC minting on Mezo by moving the
tBTC Bridge ledger there.

The RFC-3 focuses on bridging non-tBTC assets from Ethereum to Mezo.
Specifically, it aims to enable bridging of all non-Bitcoin assets
locked currently in the Portal contract on Ethereum, such as USDC,
USDT, thUSD, crvUSD, DAI, and others.

An important requirement for the proposal is to minimize the liquidity
fragmentation for major assets by establishing clear canonical token
addresses.