Add `rpcclient.Client` type assertion to `clientCtx.Client` where applicable
So far, the Cosmos SDK's `clientCtx.Client` was pointed to `rpcclient.Client`
interface directly. However, in Cosmos SDK 0.47.x a new `TendermintRPC`
interface is used for `clientCtx.Client`. The `TendermintRPC` is just
a subset of the `rpcclient.Client` interface yet the actual implementation
injected as `clientCtx.Client` still fully implements `rpcclient.Client`.
As some of the Mezo code still requires to use `rpcclient.Client`, we need
to type assert in relevant places.