Fix `NewAccountKeeper` call
The `NewAccountKeeper` does no longer accept the deprecated `AccountI`
interface from the `x/auth` module. Instead, it wants the `sdk.AccountI`
directly. Here we adjust the code accordingly.

By the way, we are restoring the `KVStoreService` parameter.
This change was reverted by mistake during a conflicting merge of
the `main` branch.