Returning a boolean in permit
This permit function returns a boolean value, differing from the permit
function in the EIP2612 standard. In the EVM, if no value is returned,
the EXTCODESIZE opcode returns 0, causing the CALL opcode to never be
executed (which would call the permit function). By returning a value,
the permit function will be correctly executed on Cosmos's underlying
blockchains.