Adjust `encoding.MakeConfig`
In Cosmos 0.50.x we need to do some additional thing within the encoding
config. First, we need to pass address codecs explicitly. Second,
we need to register custom `GetSigners` functions for messages
that doesn't use `cosmos.msg.v1.signer`. This is the case for our
`MsgEthereumTx` whose signer resolution is complex and depends on
the underlying ETH transaction.