Building domain separator in the constructor method function
Domain separator byte array depends on the chain id which is different
for testnet and mainnet. It should be constructed upon the start of the
app for the given network.
We get the chain ID from the app options map where the chain id key is
"chain-id". Tests will validate the flow from the mainnet chain id,
however it can be set to any chain id as long as the new domain
separator byte array is generated with a new chain id.