Protobuf for the x/bridge vote extension
Here we introduce a protobuf defining the `VoteExtension` type that
will be specific for the `x/bridge` module. This type will be part
of the app-level vote extension introduced in the previous commit.

The `x/bridge` vote extension holds a list of `AssetsLocked` events (emitted by
the `BitcoinBridge`contract on Ethereum) that will be collected and shared by
the Ethereum sidecar.