RFC-3: No precompiles, factory, and standard relayers
Reflecting on the idea of precompiles for canonical tokens, I realized
this idea is just terrible.

First, adding a precompile should be a hard-fork event. Are we going to
hard-fork the chain every time we want to add another canonical token?
Tokens are not what precompiles were invented for.

Second, precompile addressing overwrites EVM rules. It is not a problem
when we reserve a very specific, unique address for Bitcoin or Validator
Pool. We are writing a new code that has not been deployed anywhere
else. However, some contracts already deployed on Ethereum EVM could be
deployed on Mezo EVM under the same address - either with create2 or
using the same deployer account. The address is a derivative of the
source code. Ignoring it for not infrastructure-critical contracts feels
wrong. Not to mention, if some entity is very restrictive about their
code and addresses used by their contracts, we could enter a dangerous
territory by reusing them.

If we look at the token tracker for Ethereum, Arbitrum, and Optimism,
the token addresses are different so as long as there is a clear mapping
between addresses provided, the users should be able to navigate them.

We should avoid any special treatment of gateways or canonical tokens.
Instead, we deploy proxies and keep a central registry of token
mappings.

Last but not least, the RFC-3 states now Wormhole is the first bridging
partner and requires the use of the standard relayer to improve the user
experience.