Make `ExtendVote` returning error if sidecar returns more events than the limit
Despite the sidecar client should abide the contract defined in
the EthereumSidecarClient interface, we are doing validation of the
returned data to maintain parity with the VerifyVoteExtension logic.
The ExtendVote handler is used by honest validators so, it must
guarantee that the produced vote extension is accepted by the
VerifyVoteExtension handler.

To comply with the above, we are raising an error if the sidecar returns
more error than the `AssetsLockedEventsLimit`. Although cropping does not
break parity with VerifyVoteExtension, it can make us miss the fact
that the sidecar does not abide the contract or that the code which
calls the sidecar was in a significant way.