REJECT + error reason in x/bridge VerifyVoteExtension
In case the vote extension is invalid, we should REJECT it explicitly
and return an error describing the reason. Due to the limitations of the
Cosmos interface, REJECT without an error does not provide any details about
the reason. Conversely, error without REJECT is confusing as it should rather
denote a failure of the handler itself. The upstream app-level vote extension
handler will handle all non-ACCEPT cases gracefully and reject the app-level
vote extension.