Move commit's vote extensions signature validation to proposal sub-handlers
Originally, validation of the signatures in the app-level proposal
handler seemed to be a good move. Validation was done in one place
and sub-handlers did not have to bother about it. However, it turns
out the validation of signatures is not possible in the app-level
handler for the `ProcessProposal` hook. This is because the extended
commit info necessary to run this validation lives on the sub-handler level
and is encoded there. It is not directly available for the app-level
handler. This is a different case comparing to app-level `PrepareProposal`
where the extended commit info is directly available in the request object.
To make validation of signatures consistent across the `PrepareProposal` and
`ProcessProposal` hooks, we are moving the existing `PrepareProposal` sig
validation down, to the sub-handlers.