Custom implementation of GetStorageRoot
The intention here is to return a state root hash to comply with
https://eips.ethereum.org/EIPS/eip-7610 Proper implementation is used
to revert contract creation if address already has the non-empty storage.
However, our current codebase does not support tracking the storage root
hash that is required by the currently used go-ethereum version. For now,
we just return a dummy hash to indicate that the storage exists behind this
address. This should be good enough for now.