Properly handle app-level proposal handler failure in the VE handler
If all proposal sub-handlers fail, the composite proposal handler errors out.
As result, no app-level pseudo-tx is injected and the proposal
vector is either empty or consists only of regular app-specific
transactions.

So far, the composite VE handler handled this case inconsistently:
- If there was no regular txs in the vector and the vector was empty, it
  invoked VE sub-handlers as usual
- If regular txs were in the vector, it tried to unmarshal the first tx
  as the app-level pseudo-tx and failed, leading to empty VE being produced

Here we fix that and, we align the VE handler to always invoke sub-handlers
and do not error out if the tx vector holds only regular txs and the pseudo-tx
is not present.