Making chain handle more general purpose handle
Instead of pinning this chain handle to Mezo Portal contracts, we can
create a general purpose Ethereum chain handle that can connect to any
ethereum address upon binding generation.

Code inside of the portal.go was also moved to ethereum.go as it is more
of a core functionality of chain handle and belongs to the main
structure of the package.