Reject empty proposals in app-level `ProcessProposal` handler
So far, the behavior of this handler was not consistent as
it presence of regular chain txs decided about the outcome.
If the `PrepareProposal` handler did not inject the pseudo-tx
and there was no regular chain txs, `ProcessProposal` accepted
such a proposal. Conversely, if the `PrepareProposal` handler
did not inject the pseudo-tx but there were other regular chain
txs, `ProcessProposal` rejected such a proposal as the first
tx did not unmarshal as a pseudo-tx.

Here we make this behavior consistent and always reject the proposal
if app-level pseudo-tx was not injected due to all sub-handlers
being failed.

Moreover, from now on, we also start to reject empty proposals that
were result of block limit violation.