Drafting Ethereum connection to fetch AssetsLocked events
The goal of this commit is to establish a connection with Ethereum node
and BitcoinBridge contract to get the AssetsLocked events. This commit
is just for creating an initial plumbing. Things like subscription to
events, proper cleaning of events etc will be added later. One of the
changes here is to simplify the code that is needed to fetch the events.