Run clean before the build in make build
When I `make build`, I know I want to build. Probably the code changed
or I want to make sure the client version I am running locally is the
one I want to run. Before this change, `make build` was not building
the client if it was already built even if there were changes to the
code. Now we do a `clean` as a part of the `make build`.