Add historical info fallback in `GetPubKeyByConsAddr`
From now on, if the validator is no longer in the validator set, this function
will search for the public key in the last 10 historical info entries. It's an
attempt to maximize the chance of finding the public key as
`GetPubKeyByConsAddr` is used within the function that validates the vote
extensions. If vote extensions cannot be validated, due to missing public keys,
the consensus will halt.