Add multi-stage build to Dockerfile
This commit adds multi-stage build to the Dockerfile. The build stage
is based on the official Golang image and the final image is based on
the Distroless image. The Distroless image is a minimal image that
contains only the application and its dependencies, which makes it
smaller and more secure.

Dockerfile contains 2 final stages: debug and production. The debug
stage is based on the debug variant of the Distroless image, which
includes busybox (e.g. shell, ps, etc.) and is useful for debugging
purposes. The production stage is based on the production variant of
the Distroless image, which includes only the application and its
system dependencies.