Separate building to debug and production images
This commit separates the building of the Docker
image into two steps: one for the debug image
and one for the production image. The debug image
is built with the `debug` target, while
the production image is built with the `production`
target. The debug image is tagged with
the same tags as the production image, but with
the `debug-` prefix.