Fix make install
This command stopped to work after https://github.com/mezo-org/mezod/pull/295
was merged to main. This is because the mentioned change causes
the `BUILD_ARGS=-o` value being propagated to `go install` which produces
the "flag provided but not defined: -o" error. Explicitly passing
an empty `BUILD_ARGS` to the `install` target solves the problem.