Relaying on Finalized blocks only to fetch AssetsLocked events
The main change in this commit is how we read the AssetsLocked events.

- Replaced the current approach from buffering new AssetsLocked events
for later processing when the blocks become final, to the new approach
where we fetch only finalized blocks in a given epoch. This new approach
reduces complexity of having a buffer slice for keeping new upcoming
events and then moving those events to a finalized slice when the blocks
become final.

- Exposing public functions for BlockCounter, FinalizedBlock and Client
  that are used inside the server.go

- Added validation of the slice array upon writing new events there.

- Other minor improvements and cleanups of the code