Validating only newly added events
Before adding new events to the existing events list it is enought to
validate a new batch of events, since the events in the existing list
were already checked before. When the events list reaches 500k it would
not be optimal to validate the entire list when adding just a small
batch of new events.