Remove `tx` subtree and limit `query` to the bare minimum
As we won't support Cosmos transactions, the `tx` commands subtree
is not necessary. We are removing it.

We are also limit the `query` subtree to the bare minimum and, we
disable queries for specific modules. The relevant state of
modules will be exposed using EVM precompiles.

Last but not least, we are moving the `status` command from
root under the `query` umbrella. This makes the root CLI cleaner.