Add Docker Compose deployment
This commit adds Docker Compose deployment for the Mezod node. The deployment
consists of the following services: `init`, `cli`, `mezod`, `ethereum-sidecar`,
and `connect-sidecar`. The `init` service initializes the Mezod node
configuration and keyring. The `cli` service is used to generate validator data.
The `mezod` service runs the Mezod node. The `ethereum-sidecar` service runs the
Ethereum sidecar. The `connect-sidecar` service runs the Connect sidecar.