Revert consensus breaking change of `ValidatorPool.EvmByteCode`
Recently, we added the new `cleanupApplications` function to the `ValidatorPool`
precompile and updated its `EvmByteCode` accordingly. However, the
`EvmByteCode` is used during the `InitChain` ABCI hook. Existing testnet
nodes still use the old value so, new nodes joining the network
and using the new value of `EvmByteCode` will end up with a different
hash of the app state. In result, new nodes would not be able to
synchronize.

Here we bring back the old value to ensure backward compatibility.
The value update must be done using the on-chain upgrade mechanism
and an appropriate state migration.