Bring back `entrypoint.sh` file
When we pulled out the validator kit to a separate repo, we did not
realize it contains the entrypoint.sh file used by the Dockerfile.
Here we bring back this file. At the same time, entrypoint.sh is not
required by the validator kit itself and will be removed from there in a
separate PR.