Integrate precompiles versioning with the execution engine
Here we integrate `precompile.VersionMap` with the mechanism that
registers custom precompiles in the EVM instance created for each
transaction.

For the validator pool on testnet mezo_31611-1, we are setting the version map
to resolve to the legacy version if the execution block is lesser than 496901.
This block contains the first `submitApplication` that consumed gas according
to the new rules hence marks the boundary between the legacy and the current
validator pool precompile.