GCP infrastructure for SAFE UI
Here we add the following GCP components to run the SAFE UI for matsnet.

First of all, we are creating a dedicated VM that will host the SAFE stack.
This VM does not have a public IP. We are using IAP-tunneling to reach
it using SSH. This requires the connecting user to have the
"IAP-secured Tunnel User" IAM role and a firewall rule allowing ingress
traffic on port 22 from IAP.

The SAFE stack running on the VM is exposed to the public internet through a
global HTTPS load balancer. The load balancer uses a static global IP
and a pre-shared SSL certificate (origin cert generated on Cloudflare).
A Cloudflare proxied DNS (safe.test.mezo.org) has been configured to point the
IP of the load balancer.

On the networking level, we leveraged the existing VPC but created a separate
subnet that is used by the VM. We also configured Cloud NAT accordingly so
the VM can reach the public internet.