Schedule the v0_3 testnet fork
The mentioned fork schedules the upgrade plan for v0_3 on Mezo testnet.
This code will run on the current v0.2.0-rc2 version and will
cause the chain halt on block 1_093_500. Validators will have to
bump their nodes to v0.3.0-rc0 in order to apply the upgrade
plan and resume consensus.