Return an empty hash from the GetStorageRoot function
During the work on the go-ethereum upgrade, we decided to return
hardcoded dummy storage roots in case the state object is not empty.
Our `Account` struct has no notion of a storage root. Our `Commit()`
function has no notion of a root hash which is present in go-ethereum.
Adding support for a proper tracking of storage roots is a non-trivial
task, hence the decision for a dummy hash.

Unfortunately, this approach makes it impossible to deploy contracts
under already funded addresses. The EVM reverts transactions with
`contract address collision` given `storageRoot != types.EmptyRootHash`
in the EVM's `create` function being `true` for already funded accounts.

To address the problem with contract creation, we changed the
`StorageRoot()` implementation to always return an empty hash. With the
v1.14.8 go-ethereum dependency, this is acceptable because:
* Our Keeper's `DeleteAccount()` function for StateDB removes both the
  code and storage in one shot. This way, deploying a new contract
  under a self-destructed address that holds a non-empty storage within
  the same transaction is not possible. Deploying a new contract will be
  possible as the nonce will be reset to 0 but the storage will be
  removed first.
* EIP-7610 was meant for certain specific cases in the old EVM state
  that are not possible to happen in Mezo. With EIP-161 the nonce is
  incremented and deploying a contract with a zero nonce and zero code
  length but with non-empty storage is not possible.

This approach is fine for now but is definitely not future-proof enough.
The work for storage root tracking has been captured in
https://github.com/mezo-org/mezod/issues/369.