Introduce the `StorageRootStrategy` parameter to x/evm
Here we add a new parameter allowing us to steer the behavior
of the `GetStorageRoot` method from the EVM StateDB component.
Such a strategy is necessary to introduce changes from
https://github.com/mezo-org/mezod/pull/370 in a backward-compatible manner.
This is because the testnet live chain already contains some failed
transactions computed according to the `DummyHash` strategy and
a simple upgrade to `EmptyHash` will cause compatibility issues
for nodes syncing with the chain using the upgraded version.