Limiting requests to an Ethereum RPC provider
Each RPC provider has it's own rules in terms of the number of requests
it's accept per second. We noticed that e.g. Alchemy accepts more than
Infura. This PR aims at providing an option to adjust how many requests
can be send by ethereum sidecar to configured RPC provider. The default
value is 10 requests per second with the default 1000 batch size results
in 52 sec for a sidecar to start.
--ethereum-sidecar.server.requests-per-minute flag can be used to modify
600 requests per minute (10requests per sec).