On-chain precompile versioning
Here we refactor the current precompiles versioning mechanism by integrating
it with the on-chain state and our broader fork/upgrade mechanism from the
app/upgrades package. Thanks to that, we can execute precompile upgrades
in a much safer and easier way.

Each precompile defines a version map that assigns the specific version
number to the appropriate instance of the `precompile.Contract`.
The x/evm manages which version of the given precompile should be used
during transaction processing. If the given version stored by the x/evm
module is not registered in the keeper, the transaction targeting the
precompile in question will be reverted.

For new chains, the x/evm module is initialized with the latest versions
of the precompiles at genesis.

For the existing Matsnet chain, the new on-chain precompiles versioning
will be enabled at the v0.5.0 fork block. However, there are 4 precompiles
that are already live on this chain: BTCToken, ValidatorPool, Maintenance, and
the AssetsBridge. To ensure those precompiles work during the interim between
the upgrade to v0.5.0 and the fork block, we are doing a special trick.
Before the v0.5.0 fork block, the x/evm module will resolve all precompile
versions to 0. The mentioned 4 precompiles register a `precompile.Contract` for
v0 in the x/evm keeper hence transactions against them during the interim will
be handled as expected. Once the fork block happens, the new versioning will
be enabled and all precompile versions will be set to the latest ones in the
x/evm keeper. That will upgrade the Maintenance precompile from v0 to v2,
and enable two new precompiles: Upgrade and PriceOracle.