Use @mezo-org/contracts@0.4.0-dev for bindings explicitly
During the ongoing ERC20 bridge work, we replaced the old BitcoinBridge contract
with a new MezoBridge contract. The artifact for the new contract was published
within @mezo-org/contracts@0.4.0-dev.1 npm package. This is now the latest
version of @mezo-org/contracts so it's used by default during `make bindings`
regardless of the actual version of `mezod`.

However, ERC20 bridge work is not completed yet so we still need to stick
to the old BitcoinBridge in the upcoming v0.5.x release line. To achieve that,
we are pointing the relevant Makefile to the previous version of
@mezo-org/contracts - 0.4.0-dev - which is the last version containing
the artifact for BitcoinBridge.

This change is appropriate ONLY FOR the release/v0.5.0 branch and WILL NOT
be forward-ported to the main branch.