Add ERC20 and BTC bridging methods to the AssetsBridge precompile
Here we expose the previously implemented x/bridge CRUD methods
through the AssetsBridge precompile. This way, the PoA owner can
create/delete ERC20 mappings and read the relevant state of the bridge.

As the write methods are consensus breaking and previous version
of mezod won't understand them, we are adding those changes as a
new version of the AssetsBridge precompile - v2.