Add a CI workflow for building and publishing client binary to GCP AR
We're adding a workflow for building a client binary.
The workflow will be triggered by published tags matching the release and
release candidate format.

If release candidate tag (containing `-rc<X>` suffix) is published, the workflow
will build the client binary and publish it to
`mezo-test-420708/us-central1/mezo-staging-binary-public-tmp/mezod/ci-push-client-binary-to-gcp`
Google Cloud artifact registry.

If release tag (without the `-rc<X>` suffix) is published, the workflow will
build the client binary, compute the sha256 checksum for it and will
create a draft GitHub release with the binary and checksum added as release
attatchments.