Change output format of web3_clientVersion
So far, the web3_clientVersion RPC function returned a JSON with relevant
version information. It turns out this format causes issues in the
OpenZeppelin plugin we use to deploy upgradeable contracts behind
transparent proxies. This is because the plugin uses a helper
isDeveloperNetwork function that fetches the client name using
web3_clientVersion and assumes it's a string delimited with slashes.
In order to adhere to this convention, we change the
web3_clientVersion output format to a string
Mezod/{version}/{go_architecture}/{go_version}.