Go format configuration in pre-commit
It is now required to have the Go files formatted when committing. The solution
is less elegant than it could be given gofumpt tool used in the project does not
return a proper code when files do not pass the formatting check. This
functionality is planned to be added
(see https://github.com/mvdan/gofumpt/issues/114), but for now, I had to work
around it with `test -z` checking if the output of `make format` is non-empty.