Avoid memory aliasing when preparing bridging proposal
This change removes the potential issue with memory aliasing. The `event` is
a loop variable so even though its value changes, the memory location does not.
As a result, we've been passing the same memory location to the `addVote`
function. This worked fine because we were not capturing this memory location
anywhere and were just extracting values from the pointer but to make this code
more future-proof we should pass the address of the actual element from the
`AssetsLockedEvents` slice.