Require Go code to be linted and formatted in pre-commit
The `golangci-lint` config takes care of both linting and formatting and looks
that no one uses `gofumpt` separately. Instead of only requiring the Go code to
be formatted, we now require it to be both linted and formatted. There is still
a question why CI does not complain about linter warnings as the `make lint`
fails, but this is something that has to be solved separately.