Update default bytecode of the Maintenance precompile
Two new methods related to the chain fee splitter were added to the Maintenance precompile recently. Although we modified the ABI and the Solidity interface, we did not update the default bytecode of the precompile. This is not an issue for existing chains which no longer rely on this value but will be an issue for new chains that will load this outdated default bytecode upon genesis. If the bytecode doesn't match the actual Solidity code, we won't be able to verify the precompile on the block explorer. Here we fix that.