Prevent BTC minting to blocked addresses
When users bridge BTC to mezo using a blocked address as a
recipient for the funds, the x/bank method SendCoinsFromModuleToAccount
will fail, resulting in the minting and transfer failing down the line.

To prevent this behavour, when processing the AssetLocked events, we
now ignore events where a recipient is one of the blocked addresses.

Signed-off-by: Jeremy Letang <me@jeremyletang.com>