Add a simple bindings facade
Loading env-specific bindings at runtime is quite challenging and not
really needed as for now portal contracts have the same interfaces
on Ethereum Mainnet and Sepolia.
We just need to load the env-specific addresses.

Because of that, we are adding the new `portal` package acting as a facade for the bindings and pin up
the mainnet bindings under it.